Convert all UINT64 hash to uint32_t for texture and buffers.
This is now a 32 bit hash, so storing the result in UINT64 is misleading.  This change fixes all uses of texture and buffer hashes to be uint32_t, and that will also separate them from the remaining UINT64 hashes still used for shaders.

The changes are for parsing from ini file, comparison during hunting, and ShaderUsage reporting.

Tested in GTA5 and TheWitcher3.  Does not crash, but I'm not certain that it's fully functional.

+++ DirectX11/FrameAnalysis.cpp
@@ -601,7 +601,7 @@ void HackerContext::FrameAnalysisProcessTriggers(bool compute)
 	FrameAnalysisOptions new_options = FrameAnalysisOptions::INVALID;
 	struct ShaderOverride *shaderOverride;
 	struct TextureOverride *textureOverride;
-	UINT64 hash;
+	uint32_t hash;
 	UINT i;
 
 	// TODO: Trigger on texture inputs
